requirements.txt
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.10.0
apache-tvm-ffi==0.1.4
astor==0.8.1
asttokens==3.0.0
attrs==25.3.0
banks==2.2.0
beautifulsoup4==4.13.5
bitsandbytes==0.47.0
blake3==1.0.8
cachetools==5.5.2
cbor2==5.7.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.2
cohere==5.17.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
compressed-tensors==0.12.2
contourpy==1.3.3
cryptography==45.0.7
cuda-bindings==13.1.0
cuda-pathfinder==1.3.3
cuda-python==13.1.0
cupy-cuda12x==13.6.0
cut-cross-entropy==25.1.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.6.0
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
depyf==0.20.0
diffusers==0.35.1
dill==0.3.8
dirtyjson==1.0.8
discord==2.3.2
discord.py==2.6.3
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.1
email-validator==2.3.0
evaluate==0.4.5
executing==2.2.0
fastapi==0.124.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
fastavro==1.12.0
fastembed==0.7.3
fastrlock==0.8.3
filelock==3.19.1
filetype==1.2.0
flash_attn==2.8.3
flashinfer-python==0.5.3
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
gguf==0.17.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.9
hf_transfer==0.1.9
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface==0.0.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
inquirerpy==0.3.4
interegular==0.3.3
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-google-genai==2.1.10
langchain-huggingface==0.3.1
langchain-openai==0.3.32
langchain-text-splitters==0.3.9
langchain-unstructured==0.1.6
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.19
lark==1.2.2
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.0
llama-index-cli==0.5.0
llama-index-core==0.14.0
llama-index-embeddings-fastembed==0.4.1
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.0
llama-index-llms-openai==0.5.6
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-qdrant==0.8.4
llama-index-workflows==2.0.1
llama-parse==0.6.54
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.8.5
mmh3==5.2.0
model-hosting-container-standards==0.1.9
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.16.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl==4.3.2
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnxruntime==1.19.2
openai==1.102.0
openai-harmony==0.0.8
opencv-python-headless==4.12.0.88
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
ormsgpack==1.10.0
outlines_core==0.2.11
packaging==25.0
pandas==2.2.3
parso==0.8.5
partial-json-parser==0.2.1.1.post7
peft==0.13.2
pexpect==4.9.0
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.4.0
portalocker==3.2.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py_rust_stemmers==0.1.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycountry==24.6.1
pycparser==2.22
pydantic==2.12.5
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
qdrant-client==1.15.1
quotientai==0.4.10
ray==2.52.1
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.17.0
rignore==0.7.6
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
sentencepiece==0.2.1
sentry-sdk==2.47.0
setproctitle==1.3.7
setuptools==80.9.0
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.50.0
striprtf==0.0.26
supervisor==4.3.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.9.0
torchao==0.13.0
torchaudio==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
triton==3.5.0
trl==0.23.0
typeguard==4.4.4
typer==0.17.4
types-requests==2.32.4.20250809
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==0.9.31
tzdata==2025.2
unsloth==2025.9.4
unsloth_zoo==2025.9.5
unstructured-client==0.42.3
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vllm==0.12.0
watchfiles==1.1.1
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.3
xformers==0.0.32.post2
xgrammar==0.1.27
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
